+++ src/ser.rs
@@ -34,3 +34,6 @@ impl<'a, I, O> Serialize for SerializeTraitObject<'_, I, O> where
     map.end()
   }
 }
+
+/// Checks whether [`T`] implements [`erased_serde::Serialize`].
+pub const fn require_erased_serialize_impl<T: ?Sized + erased_serde::Serialize>() {}